refactor: optimize folders structure
optimize folders structure

MODULE-1-PART-3-gallery
