+++ gallery/gallery.js
@@ -1,6 +1,6 @@
 import { setExpireTimeAfterReloading } from "../utilities/token/setToken.js";
-import { redirectToLoginPage } from "../utilities/urlManipulation.js";
-import { tokenExists } from "../utilities/urlManipulation.js";
+import { redirectToLoginPage } from "../utilities/urlManipulation/urlManipulation.js";
+import { tokenExists } from "../utilities/urlManipulation/urlManipulation.js";
 import { renderGalleryPage, reRenderGalleryPage } from "../utilities/render/render.js";
 
 setExpireTimeAfterReloading();
